test/serenity-bdd/requirements.txt
# Dependencias para Integración SerenityBDD con Python

# Framework de pruebas BDD
pytest>=7.4.0
pytest-bdd>=6.1.0

# Reportes estilo SerenityBDD
allure-pytest>=2.13.0

# Automatización de navegador
selenium>=4.15.0
webdriver-manager>=4.0.0

# Utilidades adicionales
pytest-html>=4.0.0
pytest-xdist>=3.3.0  # Para ejecución paralela
pytest-timeout>=2.1.0

# Assertions mejoradas
assertpy>=1.1

# Para integración con el proyecto existente
requests>=2.31.0
flask>=3.0.0
